requirements.txt
affine==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.4.1
opencv-python==4.13.0.90
pillow==12.1.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.1
python-multipart==0.0.22
rasterio==1.4.4
scikit-learn==1.8.0
scipy==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
